@@ -2,7 +2,7 @@
   "name": "shopify-foundation-theme",
   "description": "Modern Shopify theme using Shopify Theme Lab, Liquid, Vue and Tailwind CSS",
   "author": "Sergej Samsonenko <contact@sergej.codes>",
-  "version": "1.2.5",
+  "version": "2.0.0",
   "license": "MIT",
   "config": {
     "reloader": {
@@ -14,12 +14,12 @@
   },
   "scripts": {
     "start": "npm-run-all --parallel --silent --race dev reloader shopify:watch",
-    "dev": "cross-env NODE_ENV=development BROWSERSLIST_ENV=development BROWSERSLIST_CONFIG=.config/.browserslistrc node_modules/.bin/webpack --config .config/webpack/webpack.dev.js --watch",
-    "build": "cross-env NODE_ENV=production BROWSERSLIST_ENV=production BROWSERSLIST_CONFIG=.config/.browserslistrc node_modules/.bin/webpack --config .config/webpack/webpack.prod.js",
+    "dev": "cross-env NODE_ENV=development BROWSERSLIST_ENV=development BROWSERSLIST_CONFIG=.config/.browserslistrc webpack --config .config/webpack/webpack.dev.js --watch",
+    "build": "cross-env NODE_ENV=production BROWSERSLIST_ENV=production BROWSERSLIST_CONFIG=.config/.browserslistrc webpack --config .config/webpack/webpack.prod.js",
     "reloader": "node node_modules/shopify-theme-lab-reloader",
-    "lint": "npm-run-all lint:*",
-    "lint:js": "node_modules/.bin/eslint src/**/*.{js,vue} --config .config/.eslintrc.js",
-    "lint:css": "node_modules/.bin/stylelint src/**/*.{vue,css,sass,scss} --config .config/.stylelintrc.js",
+    "lint": "npm-run-all lint:* --continue-on-error",
+    "lint:js": "eslint src/**/*.{js,vue} --config .config/.eslintrc.js",
+    "lint:css": "stylelint src/**/*.{vue,css,sass,scss} --config .config/.stylelintrc.js",
     "shopify:watch": "cross-env-shell shopify-themekit watch --env=dev --allow-live --config .config/shopify/shopify.dev.yml --notify=http://localhost:$npm_package_config_reloader_serverPort/reload",
     "shopify:init": "npx themelab shopify:init",
     "shopify:themes": "shopify-themekit get --list",
@@ -32,25 +32,24 @@
   },
   "dependencies": {
     "axios": "^0.21.1",
-    "tailwindcss": "^2.1.2",
-    "v-lazy-image": "^1.4.0",
-    "vue": "^2.6.12",
-    "vue-screen": "^1.5.4",
-    "vuex": "^3.6.2"
+    "tailwindcss": "^2.1.4",
+    "vue": "^3.1.1",
+    "vuex": "^4.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.14.3",
     "@babel/plugin-transform-runtime": "^7.14.3",
-    "@babel/preset-env": "^7.14.2",
-    "@shopify/themekit": "^1.1.7",
-    "autoprefixer": "^10.2.5",
+    "@babel/preset-env": "^7.14.4",
+    "@shopify/themekit": "^1.1.8",
+    "@vue/compiler-sfc": "^3.1.1",
+    "autoprefixer": "^10.2.6",
     "babel-loader": "^8.2.2",
     "clean-webpack-plugin": "^3.0.0",
     "cross-env": "^7.0.3",
-    "css-loader": "^5.2.5",
-    "css-minimizer-webpack-plugin": "^3.0.0",
-    "eslint": "^7.27.0",
-    "eslint-plugin-vue": "^7.9.0",
+    "css-loader": "^5.2.6",
+    "css-minimizer-webpack-plugin": "^3.0.1",
+    "eslint": "^7.28.0",
+    "eslint-plugin-vue": "^7.10.0",
     "eslint-webpack-plugin": "^2.5.4",
     "mini-css-extract-plugin": "^1.6.0",
     "npm-run-all": "^4.1.5",
@@ -65,10 +64,9 @@
     "stylelint-config-recommended": "^5.0.0",
     "stylelint-webpack-plugin": "^2.1.1",
     "url-loader": "^4.1.1",
-    "vue-loader": "^15.9.7",
-    "vue-template-compiler": "^2.6.12",
-    "webpack": "^5.37.1",
-    "webpack-cli": "^4.7.0",
-    "webpack-merge": "^5.7.3"
+    "vue-loader": "^16.2.0",
+    "webpack": "^5.38.1",
+    "webpack-cli": "^4.7.2",
+    "webpack-merge": "^5.8.0"
   }
 }
