@@ -1,7 +1,7 @@
 export default {
-  name: 'select-option',
+  name: 'selectOption',
   directive: {
-    inserted ($el, binding) {
+    beforeMount ($el, binding) {
       const result = Array.from($el).find(el => el.text === binding.value)
       if (result) result.selected = true
     }
